OvmfPkg/BmDmaLib: Add SEV support
When SEV is enabled, the DMA operations must be performed on a shared
(i.e unencrypted) pages. The patch adds SEV specific hooks to use the
bounce buffer when caller map/unmap host address to a DMA address and
similarly clears/set memory encryption attribute when caller allocates or
free the DMA pages.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>