OvmfPkg/VirtioNetDxe: map transmit buffer host address to device address
Update VirtioNetTransmit() to map the callers transmit buffer to a device
address and use the DeviceAddress in vring descriptors.

Since the transmit buffers are returned back to caller in
VirtioNetGetStatus() hence we maintain a simple list to map DeviceAddress
to HostAddress. The list is used for mapping to HostAddress after the
data is succesfully transmitted.

Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Jordan Justen <jordan.l.justen@intel.com>
Cc: Tom Lendacky <thomas.lendacky@amd.com>
Cc: Laszlo Ersek <lersek@redhat.com>
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>