OvmfPkg/BaseMemEncryptLib: Provide support to issue unencrypted hypercall
By default all the guest memory regions are considered encrypted, if a guest
changes the encryption attribute of the page (e.g mark a page as decrypted)
then notify hypervisor. Hypervisor will need to track the unencrypted pages
The information will be used during guest migration. If a guest memory region
is encrypted with guest key then during transmit we re-encrypt it to protect
otherwise use the default method to send the page.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>